+++ android/src/main/java/com/theoplayer/theolive/THEOliveModule.kt
@@ -54,7 +54,7 @@ class THEOliveModule(context: ReactApplicationContext) : ReactContextBaseJavaMod
   }
 
   @ReactMethod
-  fun setAuthToken(tag: Int, token: String) {
+  fun setAuthToken(tag: Int, token: String?) {
     viewResolver.resolveViewByTag(tag) { view: ReactTHEOplayerView? ->
       view?.player?.theoLive?.authToken = token
     }