+++ example/src/custom/MediaCacheDownloadButton.tsx
@@ -4,10 +4,10 @@ import { MediaCache } from 'react-native-theoplayer';
 import { MenuRadioButton } from '@theoplayer/react-native-ui/src/ui/components/menu/common/MenuRadioButton';
 
 export const MediaCacheDownloadButton = () => {
-  const playerContext = useContext(PlayerContext);
+  const { player } = useContext(PlayerContext);
 
   const onPress = useCallback(() => {
-    const currentSource = playerContext.player?.source;
+    const currentSource = player?.source;
     if (currentSource) {
       MediaCache.createTask(currentSource, {
         amount: '100%',
@@ -20,7 +20,7 @@ export const MediaCacheDownloadButton = () => {
         },
       });
     }
-  }, []);
+  }, [player]);
 
   return <MenuRadioButton label={'Create download task'} uid={0} onSelect={onPress} />;
 };