+++ ios/THEOplayerRCTBridge.m
@@ -120,6 +120,8 @@ RCT_EXTERN_METHOD(setTextTrackStyle:(nonnull NSNumber *)node
 RCT_EXTERN_METHOD(setKeepScreenOn:(nonnull NSNumber *)node
                   keepScreenOn:(BOOL)keepScreenOn)
 
+RCT_EXTERN_METHOD(willUnmount:(nonnull NSNumber *)node)
+
 RCT_EXTERN__BLOCKING_SYNCHRONOUS_METHOD(getUsableScreenDimensions)
 
 @end
@@ -301,6 +303,10 @@ RCT_EXTERN_METHOD(scheduledInterstitials:(nonnull NSNumber *)node
                   rejecter:(RCTPromiseRejectBlock)reject)
 
 RCT_EXTERN_METHOD(replaceAdTagParameters:(nonnull NSNumber *)node
+                  adTagParameters:(nullable NSDictionary *)adTagParameters)
+
+RCT_EXTERN_METHOD(setAdTagParameters:(nonnull NSNumber *)node
+                  id:(nonnull NSString *)id
                   adTagParameters:(nullable NSDictionary *)adTagParameters)
 
 @end