+++ src/internal/adapter/THEOplayerAdapter.ts
@@ -232,6 +232,10 @@ export class THEOplayerAdapter extends DefaultEventDispatcher<PlayerEventMap> im
     return this._theoAdsAdapter;
   }
 
+  get theoLive(): TheoLiveAPI {
+    return this._theoliveAdapter;
+  }
+
   get theolive(): TheoLiveAPI {
     return this._theoliveAdapter;
   }
@@ -545,5 +549,11 @@ export class THEOplayerAdapter extends DefaultEventDispatcher<PlayerEventMap> im
 
   get videoHeight(): number | undefined {
     return this._state.videoHeight;
+  }
+
+  willUnmount(): void {
+    if (Platform.OS === 'ios') {
+      NativePlayerModule.willUnmount(this._view.nativeHandle);
+    }
   }
 }