+++ src/internal/adapter/event/native/NativeTheoAdsEvent.ts
@@ -1,6 +1,7 @@
 import { Interstitial, TheoAdsEvent, TheoAdsEventType } from 'react-native-theoplayer';
 import { NativeSyntheticEvent } from 'react-native';
 import { DefaultTheoAdsErrorEvent, DefaultTheoAdsEvent } from '../PlayerEvents';
+import { createNativeInterstitial } from '../../theoads/NativeInterstitialAdapter';
 
 export interface NativeTheoAdsEvent {
   /**
@@ -19,12 +20,16 @@ export interface NativeTheoAdsEvent {
   message: string | undefined;
 }
 
-export function fromNativeTheoAdsEvent(event: NativeSyntheticEvent<NativeTheoAdsEvent>): TheoAdsEvent {
+export function fromNativeTheoAdsEvent(node: number, event: NativeSyntheticEvent<NativeTheoAdsEvent>): TheoAdsEvent {
   const { nativeEvent } = event;
   switch (event.type) {
     case TheoAdsEventType.INTERSTITIAL_ERROR:
-      return new DefaultTheoAdsErrorEvent(TheoAdsEventType.INTERSTITIAL_ERROR, nativeEvent.interstitial, nativeEvent.message);
+      return new DefaultTheoAdsErrorEvent(
+        TheoAdsEventType.INTERSTITIAL_ERROR,
+        createNativeInterstitial(node, nativeEvent.interstitial),
+        nativeEvent.message,
+      );
     default:
-      return new DefaultTheoAdsEvent(event.type as TheoAdsEventType, nativeEvent.interstitial);
+      return new DefaultTheoAdsEvent(event.type as TheoAdsEventType, createNativeInterstitial(node, nativeEvent.interstitial));
   }
 }